Fix compiler warnings across FieldData, LRUCache, SimulationConfig, tests
Silences -Wsign-conversion, -Wconversion, -Wunused-result, and
-Wdangling-else warnings surfaced by CompilerWarnings.cmake on gcc-13:

- FieldData::operator() — extract a private linear_index() helper that
  casts every operand (x, y, z, Ny_, Nz_) to std::size_t explicitly, so
  the multiplication no longer implicitly widens signed int to size_t.

- PerFaceBoundary::operator[] — cast Face enum to std::size_t directly
  (instead of int) so the array subscript no longer triggers an
  implicit sign conversion.

- LRUCache::hit_rate — explicitly cast the size_t denominator to double
  to silence -Wconversion.

- test_LRUCache — (void)-cast the return of [[nodiscard]] cache.get()
  calls whose only purpose is to promote an entry / exercise counters.

- test_ConcurrentMap — add braces around the if body inside the reader
  lambda so EXPECT_GE (which expands to if/else) cannot be misparsed
  as a dangling else.

No functional change; all 144 tests continue to pass.

+++ src/core/SimulationConfig.hpp
@@ -109,9 +109,11 @@ struct PerFaceBoundary {
 
     /// Access by face index.
     [[nodiscard]] const BoundaryConfig& operator[](Face f) const {
-        return faces[static_cast<int>(f)];
+        return faces[static_cast<std::size_t>(f)];
     }
-    [[nodiscard]] BoundaryConfig& operator[](Face f) { return faces[static_cast<int>(f)]; }
+    [[nodiscard]] BoundaryConfig& operator[](Face f) {
+        return faces[static_cast<std::size_t>(f)];
+    }
 
     /// Access by axis (0-2) and side (0=lo, 1=hi).
     [[nodiscard]] const BoundaryConfig& get(int axis, int side) const {
